+++ src/components/Navbar/SearchDrawer/SearchDrawer.tsx
@@ -3,7 +3,7 @@ import { selectNavbar, setIsSearchDrawerOpen } from 'src/redux/slices/navbar';
 import { useDispatch, useSelector } from 'react-redux';
 import { useRouter } from 'next/router';
 import useTranslation from 'next-translate/useTranslation';
-import Button, { ButtonSize, ButtonVariant } from 'src/components/dls/ButtonNew/ButtonNew';
+import Button, { ButtonSize, ButtonVariant } from 'src/components/dls/Button/Button';
 import useElementComputedPropertyValue from 'src/hooks/useElementComputedPropertyValue';
 import { getSearchResults } from 'src/api';
 import { SearchResponse } from 'types/APIResponses';
