+++ src/components/Verse/PlayVerseAudioButton.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import { playFrom } from 'src/redux/slices/AudioPlayer/state';
 
 import PlayIcon from '../../../public/icons/play-circle-outline.svg';
-import Button, { ButtonVariant } from '../dls/ButtonNew/ButtonNew';
+import Button, { ButtonVariant } from '../dls/Button/Button';
 
 interface PlayVerseAudioProps {
   timestamp: number;
